Initial Commit
Setting up the base for trying to get Bedrock Level reading working! This will probably be the Bedrock version of my MCRegionJS/Gamedata-Parser project repos, which are for Legacy Console Edition.

Once I myself have a more-stable understanding of how each of the formats for the different versions work (in each of the separate test projects), then I will start working on making a parallel implementation that supports all of them, together in one place. That's in terms of the direct API, at least. So, say you would do something like `Region.read(buffer as Uint8Array, { platform: "bedrock" })`, or something like that. It would be a singular API to work on all of the different types, and it would be built with different options to read the data accordingly. You wouldn't have to worry about how to read the data yourself, you would just work with the API, and it will handle the underlying platform implementation.

This will be the base to then later making a converter between the different formats, once I have things for each of them, specced out. By then, I will have my own TypeScript definitions for the NBT data structures, and I will be edit it directly as normal JavaScript, because that's exactly what it is inside of my parsing format (NBTify, and eventually the all-in-one library for working with the different world formats).

Then a big abstraction on top of that could be to simply read 'x' world file of type `"java"`, then simply write it out to the `"bedrock"` type. Having that working, you could then simply embed that in your program with a user interface, and you'd have world conversion right there! That's the hypothetical/hopeful end goal. It may change along the way as being the end result, but that's the plan that's helping me make the decision process of how to design all of this along the way.

Working out the format for Legacy Console Edition has been pretty difficult so far because of the lack of documentation for how it works, so I thought I'd look into Bedrock to start something new, a new project to start with! As I've mentioned previously too, getting your eyes on a new project for a while can help you out with your other current projects in the works too, and that has proven true for me multiple times before. Definitely recommend if you're feeling stuck, confused, or lost :)

Edit: Oh yeah! Re-discovered this package while looking into ZLIB-LevelDB implementations for Node. I think I found it a few months ago, about. I didn't try using it yet, but it sounds to be the closest to Mojang's LevelDB fork, for Node at least. Not sure how I can get this to work on the web yet, but I'm gonna see if I can open the format in the first place, then I'll figure out a way that works on Node and the web once that's established :)

https://github.com/extremeheat/node-leveldb-zlib