Database Iterator + Key Visualizer
Looked into the type definitions for `LevelDB`, and turns out you can use an async generator iterator over the db instance itself, and it will return the keys for the database. You don't have to manually create your own iterator variable. Neat, I like that!

Wasn't sure if you have to manually call `db.close()` or not, so I added it to test if there are any differences with not calling it yourself. From what it looks like, I think you *aren't* required to call it yourself.

Mapped each of the keys to stringified versions of the buffer, just to test seeing what's in the data a little easier. I don't know much about the format yet, so I'm using that to try and learn about what is stored where.

Never been so happy to see `minecraft:chicken` as a result 😂