DB fromEntries
Woah! Big find here. Getting an item from the database returns the name of the item, and it's value in a key-value entries pair. So iterating over all of the keys in the database returns an entries-like data structure, which can be transformed into a key-value object which looks just like a regularly-transversable NBT structure, or anything similar to that. So, then if I read the `~local_player` entry as NBT too, then I simply just have a fully transversable object for all of the data. My brain kind of exploded thinking that was really cool, haha. I think the non-UTF8 keys have to be decoded into values for something else.

(These notes were from our Discord group chat, added them as-is into the commit message since they were from when I realized all of this)