Module Relative Path
Couldn't figure out why this wasn't working before. Turns out it was because the `URL` constructor returns a path that's URI escaped, and it starts with the `file://` prefix, which appears that LevelDB-ZLIB wasn't being a fan of. I think I ran into this with something else Node-related too, so it happens for a fair amount of Node packages that don't expect more-complex URLs.