TS Enum Key
I tried using a wrapper function to get the reverse key-value mapping for the `KEY` object I made, and I realized that using a simple TypeScript `enum` here is actually just what I'm looking for! I don't want to rely on this functionality, since it doesn't have a specific JavaScript equivalent, but it essentially allows exactly what I'm looking to be able to do. In this instance, having reverse mapping allows you to go both directions with the key-value pairs, and that's what I need in this special case.