Database Key Formatting
Added parsing support to the database keys! This helps with being able to see what each entry is.

I tried loading this project over on macOS, but cmake couldn't find zlib for leveldb-zlib.. and xcode-select --install didn't help fix it either :(

This is the biggest way of seeing all of the data so far! You can see the key names and types for all of the database entries, and you can also see the NBT of all files that are just plain NBT :)