Zlib Snappy DB
This is just NOT WORKING!!! uugh. I want to use either `level` or `browser-level` to open the Snappy / zlib compressed block LevelDB database, and it's just not working. This is another attempt at trying to get it to work.

Found this issue as an example, it kind of partially worked. Now it's just a totally different API in order to iterate over all of the keys. That's been the most frustrating part of all this library soup, none of the packages or versions have the same features or APIs to accomplish the ones that they do align with. I asked ChatGPT to try and help me out here, and this is very similar to the 'halucigenic' route that it tried to give me, as that was also hard to try and decyper what it was talking about.

https://github.com/Level/level/issues/98