"Chunk" Object Grouping
Realized it may be a little nicer to view Bedrock database Chunk entries as a group of 'tags', as I can probably find a way to build an NBT-looking object out of the values. That will work with making things implemented cross-platform, as I need to make the data itself usable as something that all other platforms can convert to.

yadayada, intermediate format something something

Flamingosis! flute salad