Block Entity Actor Storage
Extracted one of the Block Entity entries from the DB, to examine what it's data format looks like. From how it explains on the Wiki, it may be an arbitrary group of NBT Compound tags, which I have to figure out where each one starts, and the other one ends. This isn't a concept I've had to use with NBTify before, so I don't currently have an easy way to find out how much of a buffer a given NBT chunk takes up. Right now turning off strict mode works, because the remaining data appears to be other NBT tags, and skipping them parses the first one successfully. So I just need to figure out where each one is, I'm not sure if that information is stored somewhere in the actor chunk itself, hence why I'm trying to analyze it's contents some more, because I'm hoping there will be a map of where each of these are stored in the data, but I feel like my luck will be that there aren't any headers for things like that.