Block Entity Actor Trailing Data
Doesn't appear to be NBT, so I think we're onto something!

So it's a group of NBT compound tags for an arbitrary length (I'm not saying List because it's not a List tag, to make that distinction), then some trailing data I'm yet to figure out.

The console is logging the outputs/trailing data for 3 actors.
(Actors are LevelDB entries, to clarify; trying to get used to MC's own terminology)

The demo files newly added to `./test` are actor #2 from the console, parsed out/saved locally.