Map Predicate Extraction
Started working on the new code implementation for parsing this data, and realized that I should extract the implementation into it's own function now. Initially I tried doing everything just by simple procedural steps, now it's a bit more narrow as to what's happening when you take out the logic from the demo code.

The indentation is weird in favor of the simple diffing.