Block Entity Chest Data
Parsed out the data of the Block Entity for the Chunk, and got the complex Chest data out of it. Very cool info to learn there! Used NBTify (cli) and Dovetail to convert the file data back to binary NBT again, since I used SNBT to debug which trees were actually the data I was looking for, because my parsing logic is still very scuffed at the moment. But hey, it worked out afterall!