Changed Demo File
Changed the local demo file to be the one I am trying to parse in the script as well. And from some help with @Bayugoon, we realized that there doesn't appear to be any trailing data outside of the Compound tags, and I think I was just parsing it incorrectly.

I set up some extra tests here to figure out if I was missing anything, and looks like this current demo script is catching everything correctly. It should be reading 7 Compound tags in this test, from what it found. Now I want to rewrite the logic so it doesn't need to iterate over every single byte in the file to find if it located a Compound or not, because that's slow, and it also catches tags that are children of the main tags too, leading to false positives. That's why I had to filter out the current results by checking the status of the `id` tag property.

Devin Townsend "Hear Me" is so freaking awesome, I love it. It pairs very well with "Factions" too, also one of my favorites. The new Steven Wilson Christmas song is outstanding as well.