Format Loop Code
This is really nice now!

Here's an explanation I wrote about how the function goes through the buffer:

The control flow is that it tried parsing the file outright, asserting that it read to the end of the file
If that fails, then it gets the error byte offset from the error message (I am going to make better DX for this from NBTify, in relation to strict mode, this should be accessible as a simple property when parsing a file), then it reads from the current position until it completes, not asserting the end length of the file (non-strict mode), then it trims the view to exclude this read portion of the buffer

So I think the takeaway from this is that I can better expose NBTify's use of strict mode through the NBTData object.

https://github.com/Offroaders123/NBTify/issues/31