Back to DB Debugging
Now I'm working on moving this into the actual DB implementation, out of the demo file/script.

https://minecraft.wiki/w/Bedrock_Edition_level_format#Chunk_key_format
https://minecraft.wiki/w/Bedrock_Edition_level_format/Other_data_format (Will definitely be of help soon, didn't know this one had lots of great documentation!)

This move adds support for parsing the NBT of BlockEntity, Entity, PendingTicks, and RandomTicks actor storage data. They all use the same format, which is a group of NBT compound tags, stored adjacent to each other in the buffer.

I started to see how this could all be handled very nicely against an organized `switch` statement against the `Key.type`, and each can get handled/parsed as separate functions/implementations. My previous/current handling is just trying to parse things as it can, but it isn't detecting things on the fly like it could. Once the format is laid out in terms of what key is expected to be what, it's not too hard at all to parse each one as it should be.

That will make things simpler to hook up to Region-Types as well, as I just need to make the associations between the types returned by the NBTData parsings, which is known ahead of time as to what it can be. Of course there could be some weird broken data in there, or something could be a modded entity maybe, but things like that, it essentially follows a specific structure, that of which is what is described on the Wiki.

I think way later down the line, I will likely try implementing whatever this big thing is, over in Rust or C, I just need to know how it works before I try to optimize it all big like that. I think Theo is right, languages like this are great for experimenting and finding things out, and lower-level languages are amazing for optimization and better implementations later. I've learned a lot from the small code I've written just the last day or so, and it all came about really quickly. I really like the ergonomics here, now that it feels like a home to me. I think most languages probably feel like that to everyone, with time. Now it feels less about the language I'm writing in, and more about my own code, and how it works to get what I see wanting to be able to do. That's the same with music as well, it just takes time to get comfortable with your tools, and comfortable with your decisions and choices to get to that end result.

(Listening to the "The Flood" while writing that last one.)

Listening to Scenes from the Flood. I think it was my Devin Townsend/Gene Hoglan listening which made me think about that project, sometimes part of Devin's sound reminds me of 'flood' and that big soundscape prog vibe.