LevelDB-Zlib Work
Looking again into the API and types for leveldb-zlib, to make things a little more strict, and figure out way to make the code a little easier to use, and performant. Not much to change here, the things I tried weren't too beneficial over this current version yet. Was hoping to change the use of `Buffer` over to the plain `Uint8Array`, but `readKey()` still uses the methods of `Buffer`, so I'll try that again sometime once I get to reworking this part.

https://github.com/extremeheat/node-leveldb-zlib