BIG LEVEL PARSING
Made a huge ton of progress on parsing values from the DB! Have been working all day pretty much, so I just kept going with it, into the night as well. Already had to charge my computer again, it was 100% this morning, turn off on me when I forgot it was so low, that was around 9.

Yeah so I went bonkers with this one, a little out the window. A bunch of stuff has changed. I basically went all in on associating what values should be parsed as, with the specific type it's stored with. That was what I realized today I think, that it's not too much different of a parsing algorithm structure than to that of `NBTReader.readTag()`, where you can do a `switch` against all of the `case`s to validate/verify against, and handle each one specific to the criteria it needs. Then you return those results out of that discerner function. Next I can organize things a bit from here, to make it more tangible and organized, and also add return types, as I want to narrow the surface area of where types are used and defined, which has helped a lot with my code as of lately. If you check it in more places, the errors can't leak out as far because you trapped them in the scope of the function block you are writing.

I got parsing for most of the different types added, I'm very happy to learn that a lot of them are just plain NBT data, so it's actually not too bad of a deal to parse each one. It's just how they are stored in the database is what you have to refactor to look and work a bit easier on your own end. The ones that still return plain `Buffer` values are the ones I still need to figure out how to parse. It's more a matter of just looking at the docs and figuring it out. It was more of a bonus that so many of the other files just use NBT, yeah that was a good thing.

So diffing was less of a concern for this one, it pretty much just rehashed the whole project all over, so I wasn't too concerned about the different stages and everything.

I also haven't completed the parsing for the `SuffixKey`-based values, hence why I keep around the `key` property when `readKey()` handles it. That is a placeholder just so I can debug the data when I come back to work on this again.

https://minecraft.wiki/w/Bedrock_Edition_level_format#Chunk_key_format
https://minecraft.wiki/w/Bedrock_Edition_level_format/Other_data_format
https://stackoverflow.com/questions/75108373/how-to-check-if-a-node-js-buffer-contains-valid-utf-8 (Could use this, didn't quite apply here but is nice to know about)
https://wiki.vg/Bedrock_Edition_level_format
https://i.imgur.com/5ljYxry_d.webp
https://github.com/mmccoo/minecraft_mmccoo/blob/master/parse_bedrock.cpp
https://learn.microsoft.com/en-us/minecraft/creator/documents/actorstorage
https://www.reddit.com/r/DevinTownsend/comments/b7i98h/singularity_one_of_dts_best/ (YES)

Tonight, we (MM) were discussing the NBT spec's use of the Java UTF-8 version, which is called MUTF-8, or Modified UTF-8.
Found this JS implementation, which I will try to use for inspiration to add support for NBTify, as it is inherent that strings will serialize correctly in the same fashion to how they do in the base game as well.
https://github.com/sciencesakura/mutf-8
https://stackoverflow.com/questions/7921016/what-does-it-mean-to-say-java-modified-utf-8-encoding
https://en.wikipedia.org/wiki/UTF-8#Modified_UTF-8

https://github.com/Offroaders123/NBTify/issues/42 (NBTify hook :P)