WorldKey/SuffixKey Value Types
I didn't yet find any documentation for the NBT of map entries, going to have to look into that one myself.

These types will eventually be part of Region-Types rather than this project. And eventually, I'd rather everything maybe also be inside it's own project scope, but we'll see. Should each platform have it's own types and implementations? Or should it all be in a single monorepo, with separate folders/modules to handle everything? Or should each feature be it's own repo, and the general data can be it's own project by which everything else uses? I think I like that last one the best, it's similar to the NBTify and Region-Types setup that is currently around, each doesn't need the other to function, instead they would rather work together in tandem to get things done better. So I'm not sure how that works for the intermediate format or world format parsing, but yeah.

https://minecraft.wiki/w/Bedrock_Edition_level_format/Other_data_format