Re-Chunk Entries List
This is a re-take on merging the Chunk actor entries into grouped 'Chunk' objects again, now that the parsing for each is specifically handled.

Now I need to make associations between each Chunk key, and it's value type, since right now these aren't linked together at the type-level. Everything is just a mishmash set of unions for all of the types of keys and value types.