NBTList Parser Generic Naming
Realized I forgot to change over the naming of these when I generalized the function for parsing multiple NBT entries from a single file. It was originally just for the Block Entity actor format, but there are other actors that also use the same format, so I smoothed it out to more likely work with those too.