Temp New DB Exploring
Oh yeah, forgot I need to add support for other Structures, as well as Maps.

This demo now finds the data for TickingArea exploration, End/Nether dimension listings, some sort of tick handling entries (not sure which of the tick ones they are), Nether, and End Portal entries. Lots to explore here, and I just need to figure out what the remaining unknown tags are, so I can sort them into chunk-like objects, as well as for which dimensions they are in. I think that's the main issue with how things are structured now, and why so many of them aren't parsing correctly, because the entries are for different dimensions as well now, so my key parsing isn't up for handling that just yet. Let's see!