Multi-Dimension Key Debugging
Realized my parsing is whack for this, I think I figured out how to organize this. Found some more resources to describe the difference between dimensions as well.

I was going to continue debugging this in this commit, but things are getting hairy, so I think restructuring a little bit of the parsing can open up room for the new dimension handling.

Actually just found a very helpful resource!
https://github.com/EngineersBox/MC-BE-LevelDB-Key-Calculator
I don't know enough about Go to know how it works sadly, I wish I could implement this on my own as well

This is another one I forgot about
I'm going to use this to figure it out as well
https://wiki.vg/Bedrock_Edition_level_format#Key_derivation