Multi-Dimension Chunk Parsing!
Wow, awesome!!! Now I can parse all of the entries into an object, which represents the whole world. Not that it should be worked with in that way, or that it should be stored outside of the database that way, but rather that it's possible to view all of the data however you need to, outstanding!!!

Now I just need to add any missing values for these that aren't handled yet, as well as the types for NBT structures that aren't yet documented, as well as parsing for keys that aren't implemented yet either, like `Data3D` and ones like that.