Portals, Nether, TheEnd, TickingArea
This is getting very solid now! Now I just need to add key parsing support for `SubChunkPrefix`, since that has an additional byte at the end of the key, which has to be persisted through `SuffixKey`.

Oh, and I need to find out of any other keys which aren't yet supported here, alonside `Map` which I haven't added yet as well.

I think parsing this with a variety of worlds should open up on providing whether it's fully working or not yet. With a new throw statement I also added here, I think it should visibly not accept when it encounters an unknown key, which is good for development purposes.

Forgot to uncomment that line in `database.ts`, it wasn't enabled because it was a lot to scroll through when it outputs all of the data at once.

Oh and I need to finish support of `SuffixKey` parsing in general, come to think of it.