Map Support
Something was wrong with the RegExp expression, it wasn't matching the key properly. I'm not sure whether I need to use this version as well as the non-working match, but I'll stick with this one for now, because it does allow it to work for my own tests.

FoxyNoTail is definitely right about the world storing a bunch of extra maps in the database; Think I may have figured out why.

Looks like when you scale up a map, it keeps reference to the original one as well, with it being the parent map ID. So if you remove one of the grandparent map IDs, that's why the game crashes.

I wonder if technically you could re-parent a map to be itself, like the root map is, allowing you to remove the grandparent ones and not have the game crash, because it won't try looking for another map it used to reference.

So, you make a new Map; that's ID #1. You scale it up one level to make it a bigger view, that's ID #2, which references back to ID #1. This scales all the way up to Map ID #5, which goes back to ID #4, and so on.

It's like the IDs are a linked list, and without the other maps, you can't build the list out, hence why loading the map will crash the game.

I'm guessing on that, but however the game manages that is probably why it crashes, because it's looking for a reference to something that isn't there.