EntryKeyNameMap Types
Going to distinguish between each part of the database parsing process, since right now each of the uses are jumbled up together.

I want one function for reading plain-string based keys, which is for things like `~local_player` and `portals`.

Then the next function is for string-based ones that have trailing additional uses, maybe these are meant to be grouped into categories which I haven't quantified into categories yet. This would count for things like `actorprefix`, `map_`, and `VILLAGE_` entries. Part of the key includes data about the key as well, not just the names. So it's like the key is coupled to where the instancing should be handled. For example I should probably make `tickingarea_`-based keys into an array/map when representing the different instances isn't very helpful in a whole list of the database entries. So the key in the map would be the UUID, and it would be in a `tickingarea: {}` object.

Ok, and for the last ones, I'm already partially handling these, but Chunk actor entries, these are mapped to their individual Chunk objects, also organized by dimension and such. Should entities also be moved into these? It may make sense in terms of representing things, as an abstraction.

I'm wondering whether the ease of use of the API needs to match how it's stored in the actual world format, or whether it should be simple to use regardless of the platform complexities. Not necessarily that it's a universal NBT-ish layer, but that it groups things together which are stored separately in favor of the file format itself. I think the game-exclusive features and such don't need to be restructured to work with the data, maybe only just how it's accessed in terms of the file format and how it works with LevelDB/Actor storage.