Dedicated Server / Realms Multiplayer
Decided to try opening a multiplayer world my friends and I played on a few years ago, to see if my current parser could work with it. It's a very big world, so it does take a long time to parse *all* of the keys into the world representation object. Doing something with this for world conversion down the line will definitely need threading and better async handling, it's a lot for it to parse all at once. This is a 495 MB world though, so it is definitely on the bigger side, so I did expect it to be slow anyways. I think the best thing would be a better way to handle working with the database, not reading all of the keys in one go. I'm doing that now in a matter of learning about they Actor entry keys and values, to work in support for those before trying to optimize things a bunch.

Ok, so I found a few keys I expected to see, which I didn't have handling for, being multiplayer 'player_' entries, as well as their associated `player_server_` values, which hold the real data for each player's NBT. These are specific to playing in multiplayer, since `~local_player` handles the rest of the use cases for playing a world. This is for Realms and the Dedicated Server / LAN or Xbox playing connections.

I found two other unknown keys I have never heard nor come across yet, and I'm not sure what they are.

There were only two entries, but it looks like it's possible for there to have been more than just these. One is called `PosTrackDB-0x${number}`, and the other is called `PositionTrackDB-LastId`. The first is a suffix-based key, where it has an instancing kind of thing, hence the `${number}` ID suffix, and the other seems to be a single key for the world, which holds reference to the most recent of those `PosTrackDB` entries. I'm not sure what the purpose of these are, but I think it has something to do with multiplayer.

Not sure if this is right, but I might have a hunch for what these mean:

```
I think I might have guessed what this is for, but I'm not sure

It might have to do with player position validation, like when you get kicked from the game for "teleporting" or something like that?
Or, "You were kicked for flying"

Like when Bedrock used to (not sure if it still does it) get laggy on Realms, and kick you for flying even though you were climbing up a hill
The position tracking was a bit laggy alongside the Realm being laggy, so it miscalculated where you were, hence it though you were flying, while you were just jumping
```

This also correlates with it being for something multiplayer-related, because I've only been testing local worlds, and you don't get kicked from your own solo world for "flying". Maybe this is a Realms-specific feature? Or just the Dedicated Server in general? Cause I think Realms uses that under the hood if I'm not mistaken. I don't think playing LAN/Xbox Live will do this either, but I'm not sure. If you join someone else's world, do you get kicked for "flying" if jumping while the connection is laggy? Or rather, something similar to that. I don't remember if that's the specific case for the buggy location tracking. I think it might be related to running the world on a 'server'-like instance, hence the Dedicated Server or Realms.

https://cran.r-project.org/web/packages/rbedrock/rbedrock.pdf
https://github.com/reedacartwright/rbedrockhttps://stackoverflow.com/questions/77312855/how-much-memory-do-javascript-objects-take-in-node-js-vs-chrome (I'm concerned it will be too heavy to load everything in memory as plain objects, but I'm not certain. Of course any too-big-world will run into this, but I'm curious whether it will happen for less complex ones as well, which wouldn't be ideal. So far Node is handling the data excellently, it just takes a while to parse and sort all of the data/buffers/NBT.)