(Legacy?) Structure Data
Next I opened a really old version of my solo survival world, from v1.0.7, which has a good deal of structure data in there as well. This is helpful for having figured out the format for structure data, which I don't remember seeing specifically documented anywhere in terms of it's NBT format. I think the tags for these thankfully appear to be very similar, at least in the older versions of the game. They are very complex and hard to understand in some parts, but I think they all use a similar baseline.

I think each of the structure shapes can be delved into a common-pattern interface, and each can implement it's own custom properties on top of that, because they do indeed follow a similar pattern for each kind of structure, with each having it's own custom NBT properties for various features and things to unlock in that structure. I still have to do this for a few other structures, as this version of my world only has the Overworld and the Nether. `bridge` appears to be entries for Nether Fortresses, originally I thought it was End Cities, but I hadn't explored The End at this point, so `dimension1` is actually the Nether.

Yeah, I still need to get things like the Woodland Mansion, Strongholds, and End Cities. There might be more, I need to look into newer versions of my world to see what I find as well.

This error-based handling for encountering unsupported DB keys is helping out exponentially! It really helps you find things you don't currently support, and add handling for them once you figure out what they are. Oop, also forgot Witch Huts, not sure if Desert Temples also have bounding boxes, nor Desert Wells.

Lots of things to try out! :D

Listened to Devin Townsend - Infinity, on the second listen right now heehee.