ActorPrefix Entity Position
Looking into parsing where actors are positioned in the world, it's just the remainder of the actor's key in the database. There's still two optional trailing bytes that I'm not sure about, it's only for some entries it seems. I thought I saw somewhere what it means, but I don't remember what that is.