Actor Storage Investigation (Digp Dilemma)
https://learn.microsoft.com/en-us/minecraft/creator/documents/actorstorage

This isn't a new find, but I think I'm connecting more things together now

```js
{
  '14, 20': 'minecraft:axolotl',
  '19, 43': 'minecraft:wandering_trader',
  '20, 16': 'minecraft:villager_v2',
  '20, 17': 'minecraft:villager_v2',
  '20, 18': 'minecraft:villager_v2',
}
```

Parsing all of the `actorprefix` entries in the database, I got these for my demo world
The keys are derived from the coordinates present in the key, and any trailing data that I haven't figured out how to parse yet I added as the third optional `<Buffer * *>` value in the mapping result
So only 2 of the entities appear to have different key formats than just the plain `actorprefix`, followed by 8 bytes for the `x, z` coordinates
I was trying to figure out where you are supposed to figure out the dimension of the entity, since the NBT doesn't store this information, having looked into it, and they key doesn't hold that info either, because the End Crystals and Striders would have to have bonus trailing data to show the dimension they were saved in too

Since they don't, I thought they must be stored some other way
Went back to look at that article, and from what I read, it seems that the mystery `digp` entries are mappings for where you locate these `x, z` entries, and they are static things to locate to, or something like that
So rather than the `x, z` Chunk of the entity being stored right in it's key and possibly changing with time for one reason or another, I think it's instead a reference to getting the associated `digp` entry, which instead holds the mapping between the Chunk and the ID that the actor/entity references in it's key

```js
{
  '14, 20': 'actorprefix\x00\x00\x00\x0E\x00\x00\x00\x14',
  '19, 43': 'actorprefix\x00\x00\x00\x13\x00\x00\x00+',
  '20, 16': 'actorprefix\x00\x00\x00\x14\x00\x00\x00\x10',
  '20, 17': 'actorprefix\x00\x00\x00\x14\x00\x00\x00\x11',
  '20, 18': 'actorprefix\x00\x00\x00\x14\x00\x00\x00\x12',
}
```

Here's a better representation of the real keys I was talking about
So this is instead showing the real keys in the database, and these are the `digp` references I was talking about
I think the `x, z` coordinates in those keys there are instead meant to reference associated `digp` keys, which are where you get the coordinates for the real chunk these were derived from

Eventually keeping track of and calculating these things for the conversion step and viewing worlds in general is going to be a bit complex, I'm not quite sure what the best way to handle this would be

Next part:
Turns out I was wrong about the `actorprefix${number}` part, the trailing numbers aren't coordinates, and are instead a UUID for the actor.

This *might* not be as difficult of a thing as I thought, but we'll see. More in a matter of being able to both parse it and replicate it when re-saving custom data to the database, and allowing it to be fully valid as to how the world format is expected to work here.

Snowcow! Works nice for Christmas, as well as The Scambot Holiday Special. Want to listen to Devin's Silent Night too.