LevelChunkMetaDataDictionary
Loaded up a new world and generated all of the available structures to be able to see what the actors for them look like, but it seems those are stored in different keys than the legacy ones I encountered the other day. So I'm on another new mission in figuring out where the NBT for structure bounding boxes are saved.

I started by looking into this one, because I saw `0a 00` after a few bytes into the file. And won't you look at that, it does have NBT data in there! It's not quite compatible with `readNBTList()` though, because each entry appears to have a little header with it, after having delved into the bytes of it with HexEd.it.

So this wasn't actually the one I was looking for, because it doesn't appear to store and structure data here, so I'm going to tangent and figure out the format of this one in the meantime, now that I've discovered it.