Key Parse Organization
Trying to split up each of the different kinds of keys into separate functions, which will make it easier to debug what's going on and what each key is.

The types are a bit wrong, and I am deciding if I want to do this by entry rather than by key. We'll see!