Proper Value Error Handling
Now the handling for unsupported values and keys are handled outside of the individual type functions.

Got stuck on the rabbit hole of watching reaction videos to Devin Townsend's 'Singularity'. 'There Be Monsters' is so great. I definitely jive with that energy, and I want to continue forward with that kind of emotion in my own songs' creativity later on.