NBT Parse Error Usage
Tried out bringing over the experimental new error object I made over in NBTify, which allows you to find out where the parsing stage failed, as well as access the data that was read thus far, and to be able to easily know where the reading process left off, allowing you to continue reading afterwards later on. This helps a lot with the `readNBTList()` implementation, because chunks of NBT data are placed adjacent to each other, and I had to do some error message-reading magic logic to get that info. This simplifies this concept, by plainly providing this error to the user.

I'm not sure how I want this to look yet, I think it is a great concept to have though! It's like how `DOMException` works in conjunction with when `showOpenFilePicker()` throws, and you can actually get the info about why it threw using that error type. I didn't know that was something you could do, and that was a big find for me!

Git Submodules are extra cool, I wish I had used them for STE with NumText and MenuDrop way back, before I used a bundler or npm. Not sure how submodules work in conjunction with GitHub Pages though. My hope would be that they are cloned when the page is built as well, so definitely something I'll have to look into.

I would have liked to just install the source from GitHub using npm, but it doesn't work with my TS code, because of the `.npmignore` file I have, it excludes the `src` directory. Otherwise, I would like to use that instead, if it did work for that. Too bad there's not a distinction to ignore the `.npmignore` when cloning a specific project. Maybe that is already some flag I don't know about?

https://github.com/Offroaders123/NBTify/commit/388ac02b0ed79e79d227c12e6ffb1195155561ef
https://github.com/Offroaders123/NBTify/issues/31
https://git-scm.com/book/en/v2/Git-Tools-Submodules
https://stackoverflow.com/questions/10914022/how-do-i-check-out-a-specific-version-of-a-submodule-using-git-submodule