Generic Map Function Strict Returns
While working on PiPe Levels, I realized what the issue I was having here was. It was because the return type of the map function is bound to the input you pass in, so it's like the parameter-based "type instantiation" concerns that it sometimes errors with as well. I don't like specifically casting the return type, since I feel like it may be less strict than a normal one. But this does indeed allow for the stricter parameter-to-return type user-accessibility that I like from an API perspective. I haven't validated it being a proper alternative/solution to this yet, but I'm also wondering if using an overload for this may work as well. I'm not completely sure though. Say if the overload was the generic mapped version, and the function's implementation type could be more broad to plain primitive/union types. Kind of like how `document.createElement()` and `document.querySelector()` are generic, until you pass in an unknown specifier, which then uses the broad return type values instead. Is that possible with a function implementation like the value ones here? I know it's possible to use that dual-typed overload setup to describe existing JS functions, but using those types to also validate the internals of how the function works is where I'm not sure if it will work, hence why it seems to be trouble to handle here.

Also discovered that I didn't add the "PositionTrackDB-LastId" entry to the `WorldKeyNameMap`.

https://github.com/Offroaders123/PiPe-Levels/blob/9ecac9f408abe0050a219fc6247511fcebb2794f/src/world.ts