DimensionID Debugging + Village Entry Discovery
Looked into the ID of the player when in other dimensions, and it is thankfully the same as the IDs used for the DB implementation. So I think my Region-Types definitions were off.

https://github.com/Offroaders123/Region-Types/issues/6

While loading the test world to change dimensions and test what value is used, I discovered that one of the more recent versions of the game uses additional Village actor keys, which I'm not completely sure about their format. I added checks to validate them with RegEx checks for the time being, not sure what their actual values are yet, and if it's just a renaming/alias of the original actor keys I'm already accounting for.